+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Removed requirement to install binaryen. The `wasm-opt` tool is now compiled into `cargo-contract`.
+
 ## [2.0.0-alpha.4] - 2022-10-03
 
 ### Fixed